--- src/utils/FormItemMap.ts
@@ -1,7 +1,0 @@
-import MemberSelect from "../components/MemberSelect";
-import { ActivityList } from "../data/ActivityImg";
-
-export const ComponentField = {
-    MEMBER_SELECT: MemberSelect,
-    ACTIVITY_SELECT: ActivityList
-};
